+++ pkg/abstract/model/transfer.go
@@ -39,7 +39,7 @@ const (
 	// Zero value is reserved and MUST NOT be used.
 	//
 	// When incrementing this value, DO ADD a link to the function(s) implementing this fallback to CHANGELOG.md in the current directory
-	LatestVersion int = 9
+	LatestVersion int = 10
 	// NewTransfersVersion is the version of the typesystem set for new transfers. It must be less or equal to the LatestVersion.
 	//
 	// To upgrade typesystem version, the following process should be applied: